refactor(perl-module): split use lib resolver modules (#8197)
