+++ crates/perl-module/src/resolution/use_lib.rs
@@ -3,7 +3,15 @@
 //! Scans Perl source text for `use lib` pragmas and recognizes common
 //! `FindBin` patterns to discover additional module include directories.
 
-use std::path::{Component, Path, PathBuf};
+use std::path::Path;
+
+mod extract;
+mod resolve;
+mod statements;
+
+use extract::extract_paths_from_args;
+pub use resolve::resolve_use_lib_paths;
+use statements::{split_perl_statements, strip_no_lib_prefix, strip_use_lib_prefix};
 
 /// A discovered include path from a `use lib` statement.
 #[derive(Debug, Clone, PartialEq, Eq)]
@@ -74,137 +82,6 @@ pub fn extract_use_lib_operations(source: &str) -> Vec<UseLibAction> {
     }
 
     ops
-}
-
-fn split_perl_statements(source: &str) -> Vec<&str> {
-    let mut statements = Vec::new();
-    let mut start = 0;
-    let mut in_single = false;
-    let mut in_double = false;
-    let mut escaped = false;
-    // Whether any non-whitespace, non-comment content has appeared in the
-    // current statement since `start`.  When false and we hit a comment, we
-    // can safely advance `start` past the comment so it doesn't pollute the
-    // next statement slice.
-    let mut has_content = false;
-
-    let chars: Vec<(usize, char)> = source.char_indices().collect();
-    let mut i = 0;
-
-    while i < chars.len() {
-        let (idx, ch) = chars[i];
-
-        if escaped {
-            escaped = false;
-            i += 1;
-            continue;
-        }
-
-        if ch == '\\' && (in_single || in_double) {
-            escaped = true;
-            i += 1;
-            continue;
-        }
-
-        if ch == '\'' && !in_double {
-            in_single = !in_single;
-            has_content = true;
-            i += 1;
-            continue;
-        }
-
-        if ch == '"' && !in_single {
-            in_double = !in_double;
-            has_content = true;
-            i += 1;
-            continue;
-        }
-
-        // Skip Perl line comments: # ... <newline>
-        // A `#` is only a comment when outside of any string literal.
-        if ch == '#' && !in_single && !in_double {
-            // Skip to end of line (or end of source).
-            let comment_end = match source[idx..].find('\n') {
-                Some(nl_offset) => idx + nl_offset + 1,
-                None => source.len(),
-            };
-            // If no statement content has been seen yet, advance `start` past
-            // the comment so the comment text is not included in the next slice.
-            if !has_content {
-                start = comment_end;
-            }
-            // Skip the iterator past the comment.
-            while i < chars.len() && chars[i].0 < comment_end {
-                i += 1;
-            }
-            continue;
-        }
-
-        if ch == ';' && !in_single && !in_double {
-            let end = idx + ch.len_utf8();
-            statements.push(&source[start..end]);
-            start = end;
-            has_content = false;
-        } else if !ch.is_whitespace() {
-            has_content = true;
-        }
-
-        i += 1;
-    }
-
-    if start < source.len() {
-        statements.push(&source[start..]);
-    }
-
-    statements
-}
-
-/// Resolve `use lib` paths against a workspace root and optional file directory.
-///
-/// - Absolute paths are accepted only when they stay under `workspace_root`.
-/// - `$FindBin::Bin`-relative paths are resolved against `file_dir` (or `workspace_root` if absent).
-/// - Other relative paths are resolved against `workspace_root`.
-pub fn resolve_use_lib_paths(
-    use_lib_paths: &[UseLibPath],
-    workspace_root: &Path,
-    file_dir: Option<&Path>,
-) -> Vec<String> {
-    let mut result = Vec::new();
-
-    for ulp in use_lib_paths {
-        let path_str = &ulp.path;
-
-        if ulp.from_findbin {
-            let base = file_dir.unwrap_or(workspace_root);
-            let Some(resolved) = normalize_findbin_path(base, path_str) else {
-                continue;
-            };
-            if resolved.strip_prefix(workspace_root).is_err() {
-                continue;
-            }
-            if let Some(s) = path_to_relative_string(&resolved, workspace_root)
-                && !result.contains(&s)
-            {
-                result.push(s);
-            }
-        } else {
-            let p = Path::new(path_str);
-            if p.is_absolute() {
-                if let Some(s) = path_to_relative_string(p, workspace_root)
-                    && !result.contains(&s)
-                {
-                    result.push(s);
-                }
-            } else {
-                let s = path_str.to_string();
-                if !result.contains(&s) {
-                    result.push(s);
-                }
-            }
-        }
-    }
-
-    result
 }
 
 /// Resolve effective include paths from lexical `use lib` / `no lib` operations.
@@ -296,232 +173,4 @@ pub fn no_lib_cancelled_paths_at_offset(
         }
     }
     cancelled
-}
-
-fn strip_use_lib_prefix(trimmed: &str) -> Option<&str> {
-    let rest = trimmed.strip_prefix("use")?;
-    if !rest.starts_with(|c: char| c.is_whitespace()) {
-        return None;
-    }
-    let rest = rest.trim_start();
-    let rest = rest.strip_prefix("lib")?;
-    if !rest.starts_with(|c: char| c.is_whitespace() || c == '(' || c == ';') {
-        return None;
-    }
-    Some(rest.trim_start())
-}
-
-fn strip_no_lib_prefix(trimmed: &str) -> Option<&str> {
-    let rest = trimmed.strip_prefix("no")?;
-    if !rest.starts_with(|c: char| c.is_whitespace()) {
-        return None;
-    }
-    let rest = rest.trim_start();
-    let rest = rest.strip_prefix("lib")?;
-    if !rest.starts_with(|c: char| c.is_whitespace() || c == '(' || c == ';') {
-        return None;
-    }
-    Some(rest.trim_start())
-}
-
-fn extract_paths_from_args(args: &str, out: &mut Vec<UseLibPath>) {
-    let args = args.trim_end_matches(';').trim();
-
-    if let Some(rest) = args.strip_prefix("qw") {
-        extract_qw_paths(rest.trim_start(), out);
-        return;
-    }
-
-    if let Some(inner) = strip_parens(args) {
-        extract_quoted_list(inner, out);
-        return;
-    }
-
-    extract_quoted_list(args, out);
-}
-
-fn extract_qw_paths(rest: &str, out: &mut Vec<UseLibPath>) {
-    let (open, close) = match rest.chars().next() {
-        Some('(') => ('(', ')'),
-        Some('/') => ('/', '/'),
-        Some('{') => ('{', '}'),
-        Some('[') => ('[', ']'),
-        Some('<') => ('<', '>'),
-        Some('!') => ('!', '!'),
-        _ => return,
-    };
-
-    let inner = &rest[open.len_utf8()..];
-    let end = inner.find(close).unwrap_or(inner.len());
-    let content = &inner[..end];
-
-    for word in content.split_whitespace() {
-        out.push(UseLibPath { path: word.to_string(), from_findbin: false });
-    }
-}
-
-fn strip_parens(s: &str) -> Option<&str> {
-    let s = s.trim();
-    let inner = s.strip_prefix('(')?;
-    let inner = inner.trim_end().strip_suffix(')')?;
-    Some(inner)
-}
-
-fn extract_quoted_list(s: &str, out: &mut Vec<UseLibPath>) {
-    let mut remaining = s.trim();
-
-    while !remaining.is_empty() {
-        remaining = remaining.trim_start_matches(|c: char| c == ',' || c.is_whitespace());
-        if remaining.is_empty() {
-            break;
-        }
-
-        // Skip Perl line comments: # ... <newline>
-        if remaining.starts_with('#') {
-            remaining = match remaining.find('\n') {
-                Some(nl) => &remaining[nl + 1..],
-                None => "",
-            };
-            continue;
-        }
-
-        if let Some((path, from_findbin, rest)) = extract_one_quoted(remaining) {
-            out.push(UseLibPath { path, from_findbin });
-            remaining = rest.trim_start_matches(|c: char| c == ',' || c.is_whitespace());
-        } else {
-            break;
-        }
-    }
-}
-
-fn extract_one_quoted(s: &str) -> Option<(String, bool, &str)> {
-    let s = s.trim();
-    let quote = match s.chars().next()? {
-        '\'' => '\'',
-        '"' => '"',
-        _ => return None,
-    };
-
-    let inner = &s[1..];
-    let end = find_unescaped_quote(inner, quote)?;
-    let content = &inner[..end];
-    let rest = &inner[end + quote.len_utf8()..];
-
-    let (path, from_findbin) = resolve_findbin_in_string(content);
-    Some((path, from_findbin, rest))
-}
-
-fn find_unescaped_quote(s: &str, quote: char) -> Option<usize> {
-    let mut escaped = false;
-
-    for (idx, ch) in s.char_indices() {
-        if escaped {
-            escaped = false;
-            continue;
-        }
-
-        if ch == '\\' {
-            escaped = true;
-            continue;
-        }
-
-        if ch == quote {
-            return Some(idx);
-        }
-    }
-
-    None
-}
-
-fn resolve_findbin_in_string(s: &str) -> (String, bool) {
-    // Fully-qualified FindBin variables — no word-boundary ambiguity because `::` terminates
-    // the name and braced forms are unambiguous.
-    let qualified_vars =
-        ["$FindBin::Bin", "$FindBin::RealBin", "${FindBin::Bin}", "${FindBin::RealBin}"];
-
-    for var in &qualified_vars {
-        if let Some(rest) = s.strip_prefix(var) {
-            let path = rest.strip_prefix('/').unwrap_or(rest);
-            if path.is_empty() {
-                return (".".to_string(), true);
-            }
-            return (path.to_string(), true);
-        }
-    }
-
-    // Short exported forms: `$Bin`, `$RealBin`, `${Bin}`, `${RealBin}`.
-    // Braced forms (`${Bin}`) are always unambiguous.  Bare forms (`$Bin`,
-    // `$RealBin`) must be followed by `/`, end-of-string, or a non-identifier
-    // character to avoid false-positives on variables like `$BinDir` or
-    // `$RealBinPath`.
-    let bare_short = ["$Bin", "$RealBin"];
-    let braced_short = ["${Bin}", "${RealBin}"];
-
-    for var in &bare_short {
-        if let Some(rest) = s.strip_prefix(var) {
-            // Word-boundary check: the character after the variable name must
-            // not be a Perl identifier character (letter, digit, or `_`).
-            // This prevents `$BinDir` or `$RealBinPath` from matching `$Bin`/`$RealBin`.
-            let next = rest.chars().next();
-            if next.is_none() || next.is_some_and(|c| !c.is_alphanumeric() && c != '_') {
-                let path = rest.strip_prefix('/').unwrap_or(rest);
-                if path.is_empty() {
-                    return (".".to_string(), true);
-                }
-                return (path.to_string(), true);
-            }
-        }
-    }
-
-    for var in &braced_short {
-        if let Some(rest) = s.strip_prefix(var) {
-            let path = rest.strip_prefix('/').unwrap_or(rest);
-            if path.is_empty() {
-                return (".".to_string(), true);
-            }
-            return (path.to_string(), true);
-        }
-    }
-
-    (s.to_string(), false)
-}
-
-fn path_to_relative_string(path: &Path, workspace_root: &Path) -> Option<String> {
-    if let Ok(rel) = path.strip_prefix(workspace_root) {
-        // Guard against lexical strip_prefix matching an embedded `..` segment.
-        // For example, `/workspace/../etc` strips the `/workspace` prefix lexically,
-        // leaving `../etc` which would escape the workspace.  Reject any result
-        // that contains a parent-directory component.
-        if rel.components().any(|c| c == std::path::Component::ParentDir) {
-            return None;
-        }
-        let s = normalize_relative_path_string(rel.to_string_lossy().as_ref());
-        if s.is_empty() { Some(".".to_string()) } else { Some(s) }
-    } else if path.is_absolute() {
-        None
-    } else {
-        let s = normalize_relative_path_string(path.to_string_lossy().as_ref());
-        Some(s)
-    }
-}
-
-fn normalize_relative_path_string(path: &str) -> String {
-    path.replace('\\', "/")
-}
-
-fn normalize_findbin_path(base: &Path, relative: &str) -> Option<PathBuf> {
-    let mut normalized = PathBuf::from(base);
-    for component in Path::new(relative).components() {
-        match component {
-            Component::CurDir => {}
-            Component::Normal(segment) => normalized.push(segment),
-            Component::ParentDir => {
-                if !normalized.pop() {
-                    return None;
-                }
-            }
-            Component::RootDir | Component::Prefix(_) => return None,
-        }
-    }
-    Some(normalized)
 }
